+++ src/lib/ares_parse_mx_reply.c
@@ -84,7 +84,7 @@ ares_parse_mx_reply (const unsigned char *abuf, int alen_int,
   aptr = abuf + HFIXEDSZ;
   status = ares__expand_name_for_response(aptr, abuf, alen, &hostname, &len, ARES_TRUE);
   if (status != ARES_SUCCESS)
-    return status;
+    return (int)status;
 
   if (aptr + len + QFIXEDSZ > abuf + alen)
     {
@@ -172,7 +172,7 @@ ares_parse_mx_reply (const unsigned char *abuf, int alen_int,
     {
       if (mx_head)
         ares_free_data (mx_head);
-      return status;
+      return (int)status;
     }
 
   /* everything looks fine, return the data */
